Move debounce effect to it's own file.
